style: clear the compiler warnings introduced in this release
Sign-conversion in VoiceHumanizer's reset loop, an implicit double-to-float
narrowing in the vowel stimulus and the NaN probe, and four float-equality
comparisons.

The equality comparisons were all deliberate - "did the caller actually change
anything" early-outs, and bit-exactness assertions where approximate equality
would defeat the point of the test - so they are expressed with
juce::exactlyEqual rather than loosened, which states the intent to the reader
as well as to -Wfloat-equal.

The two remaining -Wfloat-equal warnings in DeepDiveTests.cpp predate this
branch and are left alone.

+++ src/dsp/VoiceHumanizer.h
@@ -59,10 +59,10 @@ public:
 
     void reset() noexcept
     {
-        for (int walk = 0; walk < numWalks; ++walk)
+        for (size_t walk = 0; walk < static_cast<size_t> (numWalks); ++walk)
         {
             // Distinct, non-zero seed per walk per voice.
-            states[walk] = seedBase ^ (0xA24BAED4963EE407ull * static_cast<std::uint64_t> (walk + 1));
+            states[walk] = seedBase ^ (0xA24BAED4963EE407ull * (walk + 1));
             if (states[walk] == 0)
                 states[walk] = 0x853C49E6748FEA9Bull;
